Fixed issue with infinite battery depletion, issue with injecting new communications from previous vehicle being treated as new vehicle
Also some changes to improve code performance

+++ SimulatorBridger-IOTSimOsmosisRES/src/main/java/org/cloudbus/osmosis/core/policies/CloudletSchedulerTimeSharedMutipleCPUs.java
@@ -67,7 +67,7 @@ public class CloudletSchedulerTimeSharedMutipleCPUs extends CloudletSchedulerTim
 		}
 		getCloudletExecList().add(rcl);
 		double extraSize = getCapacity(getCurrentMipsShare()) * fileTransferTime;		
-		double length = (double) (cloudlet.getCloudletLength() + extraSize);
+		double length = cloudlet.getCloudletLength() + extraSize;
 		cloudlet.setCloudletLength(length);
 		return cloudlet.getCloudletLength() / getCapacity(getCurrentMipsShare());
 	}
